+++ changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - upgrade: update protobuf-gradle-plugin from v0.8.17 to v0.9.2
 - upgrade: update grpc from v1.49.2 to v1.50.0
 - upgrade: update protobuf from v3.21.7 to v3.21.12
+- upgrade: update docker lightninglabs/lnd from v0.15.3-beta to v0.15.5-beta
 
 ## [0.5.0] - 2022-11-29
 ### Breaking